move services_config/ to config/ section
